Specs that run under the default configuration are now parallelized. This is backwards compatible as well so the old make targets should still work.
