--- test/e2e/data_setup/config/recordadd-defaults.dev.json
@@ -1,32 +1,0 @@
-{
-    "setup": {
-        "schemaConfigurations": [{
-            "catalog": {
-                "acls": [
-                    {"name": "content_write_user", "user": "*" },
-                    {"name": "content_read_user", "user": "*" }
-                ]
-            },
-            "schema": {
-                "name": "defaults",
-                "createNew": true,
-                "path": "test/e2e/data_setup/schema/defaults.json"
-            },
-            "tables": {
-                "createNew": true
-            }
-        }],
-        "schema": "defaults"
-    },
-    "cleanup": true,
-    "tests": {
-        "params": {
-            "table_name": "defaults-table",
-            "column_names": ["text", "int"],
-            "text_value": "default",
-            "int_value": "25",
-            "boolean_true_value": "true",
-            "boolean_false_value": "false"
-        }
-    }
-}
