--- test/e2e/data_setup/config/recordadd-multi.dev.json
@@ -1,50 +1,0 @@
-{
-    "setup": {
-        "schemaConfigurations": [{
-            "catalog": {
-                "acls": [
-                    {"name": "content_write_user", "user": "*" },
-                    {"name": "content_read_user", "user": "*" }
-                ]
-            },
-            "schema": {
-                "name": "multi-add",
-                "createNew": true,
-                "path": "test/e2e/data_setup/schema/multi-add.json"
-            },
-            "tables": {
-                "createNew": true
-            },
-            "entities": {
-            	"createNew": true,
-                "path": "test/e2e/data_setup/data/multi-add"
-            }
-        }],
-        "schema": "multi-add"
-    },
-    "cleanup" : true,
-    "tests" : {
-        "params" : {
-            "multi_insert" : {
-                "table_name" : "multi-add-table",
-                "records" : 3,
-                "max_input_rows" : 200,
-                "initial_int_val" : "1",
-                "initial_text_val" : "a",
-                "input_1_text_val" : "asdfg",
-                "input_2_text_val" : "qwerty",
-                "input_3_int_val" : "17"
-            },
-            "multi_edit" : {
-                "table_name" : "multi-add-table",
-                "int_column_name" : "int",
-                "text_column_name" : "text",
-                "keys_2" : [{ "name": "id", "value": "1000", "operator" : "=" },
-                            { "name": "id", "value": "1001", "operator" : "=" }],
-                "keys_3" : [{ "name": "id", "value": "1000", "operator" : "=" },
-                            { "name": "id", "value": "1001", "operator" : "=" },
-                            { "name": "id", "value": "1002", "operator" : "=" }]
-            }
-        }
-    }
-}
