--- test/e2e/data_setup/config/recordedit-composite-key.dev.json
@@ -1,34 +1,0 @@
-{
-    "setup": {
-        "schemaConfigurations": [{
-            "catalog": {
-                "acls": [
-                    {"name": "content_write_user", "user": "*" },
-                    {"name": "content_read_user", "user": "*" }
-                ]
-            },
-            "schema": {
-                "name": "product-person",
-                "createNew": true,
-                "path": "test/e2e/data_setup/schema/product-person.json"
-            },
-            "tables": {
-            	"createNew": true
-            },
-            "entities": {
-            	"createNew": true,
-                "path": "test/e2e/data_setup/data/product"
-            }
-        }],
-        "schema": "product-person"
-    },
-    "cleanup" : true,
-    "tests" : {
-        "params": {
-            "tables" : [{
-                "table_name" : "accommodation",
-                "column_names": ["first_name", "last_name", "product-person_product-person-fk"]
-            }]
-        }
-    }
-}
