--- test/e2e/data_setup/config/recordedit-domain-filter.dev.json
@@ -1,31 +1,0 @@
-{
-    "setup": {
-        "schemaConfigurations": [{
-            "catalog": {
-                "acls": [
-                    {"name": "content_write_user", "user": "*" },
-                    {"name": "content_read_user", "user": "*" }
-                ]
-            },
-            "schema": {
-                "name": "fk-filter-pattern",
-                "createNew": true,
-                "path": "test/e2e/data_setup/schema/fk-filter-pattern.json"
-            },
-            "tables": {
-            	"createNew": true
-            },
-            "entities": {
-            	"createNew": true,
-                "path": "test/e2e/data_setup/data/fk-filter-pattern"
-            }
-        }],
-        "schema": "fk-filter-pattern"
-    },
-    "cleanup" : true,
-    "tests" : {
-        "params": {
-            "table_name": "main-entity-table"
-        }
-    }
-}
