--- test/e2e/data_setup/config/recordedit-multi-column-types.dev.json
@@ -1,84 +1,0 @@
-{
-    "setup": {
-        "schemaConfigurations": [{
-            "catalog": {
-                "acls": [
-                    {"name": "content_write_user", "user": "*" },
-                    {"name": "content_read_user", "user": "*" }
-                ]
-            },
-            "schema": {
-                "name": "multi-column-types",
-                "createNew": true,
-                "path": "test/e2e/data_setup/schema/multi-column-types.json"
-            },
-            "tables": {
-                "createNew": true
-            },
-            "entities": {
-                "createNew": true,
-                "path": "test/e2e/data_setup/data/multi-column-types"
-            }
-        }],
-        "schema": "multi-column-types"
-    },
-    "cleanup": true,
-    "tests": {
-        "params": {
-            "table_1": {
-                "tableName": "table_1",
-                "key": {"columnName": "id", "value": 1, "operator": "="},
-                "row": [
-                    {"name": "id", "value": 1, "displayType": "disabled"},
-                    {"name": "int2_null_col", "value": null, "displayType": "int2"},
-                    {"name": "int2_col", "value": 32767, "displayType": "int2"},
-                    {"name": "int4_null_col", "value": null, "displayType": "int4"},
-                    {"name": "int4_col", "value": -2147483648, "displayType": "int4"},
-                    {"name": "int8_null_col", "value": null, "displayType": "int8"},
-                    {"name": "int8_col", "value": 9007199254740991, "displayType": "int8"},
-                    {"name": "float4_null_col", "value": null, "displayType": "float4"},
-                    {"name": "float4_col", "value": 4.6123, "displayType": "float4"},
-                    {"name": "float8_null_col", "value": null, "displayType": "float8"},
-                    {"name": "float8_col", "value": 234523523.023045, "displayType": "float8"},
-                    {"name": "text_null_col", "value": null, "displayType": "text"},
-                    {"name": "text_col", "value": "sample", "displayType": "text"},
-                    {"name": "longtext_null_col", "value": null, "displayType": "longtext"},
-                    {"name": "longtext_col", "value": "asjdf;laksjdf;laj ;lkajsd;f lkajsdf;lakjs f;lakjs df;lasjd f;ladsjf;alskdjfa ;lskdjf a;lsdkjf a;lskdfjal;sdkfj as;ldfkj as;dlf kjasl;fkaj;lfkjasl;fjas;ldfkjals;dfkjas;dlkfja;sldkfjasl;dkfjas;dlfkjasl;dfkja; lsdjfk a;lskdjf a;lsdfj as;ldfja;sldkfja;lskjdfa;lskdjfa;lsdkfja;sldkfjas;ldfkjas;dlfkjas;lfkja;sldkjf a;lsjf ;laskj fa;slk jfa;sld fjas;l js;lfkajs;lfkasjf;alsja;lk ;l kja", "displayType": "longtext"},
-                    {"name": "markdown_null_col", "value": null, "displayType": "markdown"},
-                    {"name": "markdown_col", "value": "<strong>Sample</strong>", "displayType": "markdown"},
-                    {"name": "bool_null_col", "value": null, "displayType": "boolean"},
-                    {"name": "bool_true_col", "value": true, "displayType": "boolean"},
-                    {"name": "bool_false_col", "value": false, "displayType": "boolean"},
-                    {"name": "timestamp_null_col", "value": null, "displayType": "timestamp"},
-                    {"name": "timestamp_col", "value": "2016-01-18T13:00:00", "displayType": "timestamp"},
-                    {"name": "timestamptz_null_col", "value": null, "displayType": "timestamptz"},
-                    {"name": "timestamptz_col", "value": "2016-01-18T00:00:00-08:00", "displayType": "timestamptz"},
-                    {"name": "date_null_col", "value": null, "displayType": "date"},
-                    {"name": "date_col", "value": "2016-08-15", "displayType": "date"},
-                    {"name": "fk_null_col", "value": null, "displayType": "popup-select"},
-                    {"name": "fk_col", "value": "1", "displayType": "popup-select"}
-                ]
-            },
-            "table_w_generated_columns" : {
-                "tableName": "table_w_generated_columns",
-                "key": {"columnName": "id", "value": 1, "operator": "="},
-                "row": [
-                    {"name": "int2_col_gen", "value": 32767, "displayType": "disabled"},
-                    {"name": "int4_col_gen", "value": -2147483648, "displayType": "disabled"},
-                    {"name": "int8_col_gen", "value": 9007199254740991, "displayType": "disabled"},
-                    {"name": "float4_col_gen", "value": 4.6123, "displayType": "disabled"},
-                    {"name": "float8_col_gen", "value": 234523523.023045, "displayType": "disabled"},
-                    {"name": "text_col_gen", "value": "sample", "displayType": "disabled"},
-                    {"name": "longtext_col_gen", "value": "asjdf;laksjdf;laj ;lkajsd;f lkajsdf;lakjs f;lakjs df;lasjd f;ladsjf;alskdjfa ;lskdjf a;lsdkjf a;lskdfjal;sdkfj as;ldfkj as;dlf kjasl;fkaj;lfkjasl;fjas;ldfkjals;dfkjas;dlkfja;sldkfjasl;dkfjas;dlfkjasl;dfkja; lsdjfk a;lskdjf a;lsdfj as;ldfja;sldkfja;lskjdfa;lskdjfa;lsdkfja;sldkfjas;ldfkjas;dlfkjas;lfkja;sldkjf a;lsjf ;laskj fa;slk jfa;sld fjas;l js;lfkajs;lfkasjf;alsja;lk ;l kja", "displayType": "disabled"},
-                    {"name": "markdown_col_gen", "value": "<strong>Sample</strong>", "displayType": "disabled"},
-                    {"name": "bool_true_col_gen", "value": true, "displayType": "disabled"},
-                    {"name": "bool_false_col_gen", "value": false, "displayType": "disabled"},
-                    {"name": "timestamp_col_gen", "value": "2016-01-18T13:00:00", "displayType": "disabled"},
-                    {"name": "timestamptz_col_gen", "value": "2016-01-18T00:00:00-08:00", "displayType": "disabled"},
-                    {"name": "date_col_gen", "value": "2016-08-15", "displayType": "disabled"},
-                    {"name": "fk_col_gen", "value": "1", "displayType": "disabled"}
-                ]
-            }
-        }
-    }
-}
