--- test/e2e/data_setup/config/recordedit-submission-disabled.dev.json
@@ -1,34 +1,0 @@
-{
-    "setup": {
-        "schemaConfigurations": [{
-            "catalog": {
-                "acls": [
-                    {"name": "content_write_user", "user": "*" },
-                    {"name": "content_read_user", "user": "*" }
-                ]
-            },
-            "schema": {
-                "name": "product-no-serial",
-                "createNew": true,
-                "path": "test/e2e/data_setup/schema/product-no-serial.json"
-            },
-            "tables": {
-            	"createNew": true
-            },
-            "entities": {
-            	"createNew": true,
-                "path": "test/e2e/data_setup/data/product"
-            }
-        }],
-        "schema": "product-no-serial"
-    },
-    "cleanup" : true,
-    "tests" : {
-        "params": {
-            "tables" : [{
-                "table_name" : "accommodation",
-                "keys" : [{ "name": "id", "value": "2002", "operator" : "="}]
-            }]
-        }
-    }
-}
