+++ test/e2e/data_setup/config/recordedit/composite-key.config.json
@@ -6,9 +6,9 @@
         ]
     },
     "schema": {
-        "name": "recordset-product",
+        "name": "product-person",
         "createNew": true,
-        "path": "test/e2e/data_setup/schema/recordset-product.json"
+        "path": "test/e2e/data_setup/schema/product-person.json"
     },
     "tables": {
         "createNew": true
