+++ test/e2e/data_setup/config/recordedit/domain-filter.config.json
@@ -6,15 +6,15 @@
         ]
     },
     "schema": {
-        "name": "recordset-product",
+        "name": "fk-filter-pattern",
         "createNew": true,
-        "path": "test/e2e/data_setup/schema/recordset-product.json"
+        "path": "test/e2e/data_setup/schema/fk-filter-pattern.json"
     },
     "tables": {
         "createNew": true
     },
     "entities": {
         "createNew": true,
-        "path": "test/e2e/data_setup/data/product"
+        "path": "test/e2e/data_setup/data/fk-filter-pattern"
     }
 }
