+++ test/e2e/data_setup/config/recordedit/edit-multi.dev.json
@@ -1,9 +1,9 @@
 {
     "setup": {
         "schemaConfigurations": [
-            "recordset/edit.config.json"
+            "recordedit/edit-multi.config.json"
         ],
-        "schema": "recordset-multi-edit"
+        "schema": "multi-edit"
     },
     "cleanup" : true
 }
