+++ test/e2e/data_setup/config/recordedit/multi-column-types.dev.json
@@ -1,0 +1,9 @@
+{
+    "setup": {
+        "schemaConfigurations": [
+            "/recordedit/multi-column-types.config.json"
+        ],
+        "schema": "multi-column-types"
+    },
+    "cleanup": true
+}
