+++ test/e2e/data_setup/config/recordedit/submission-disabled.config.json
@@ -6,9 +6,9 @@
         ]
     },
     "schema": {
-        "name": "recordset-product",
+        "name": "product-no-serial",
         "createNew": true,
-        "path": "test/e2e/data_setup/schema/recordset-product.json"
+        "path": "test/e2e/data_setup/schema/product-no-serial.json"
     },
     "tables": {
         "createNew": true
