+++ test/e2e/data_setup/config/recordedit/submission-disabled.dev.json
@@ -1,0 +1,9 @@
+{
+    "setup": {
+        "schemaConfigurations": [
+            "recordedit/submission-disabled.config.json"
+        ],
+        "schema": "product-no-serial"
+    },
+    "cleanup" : true
+}
