+++ test/e2e/data_setup/config/recordset/add.dev.json
@@ -3,20 +3,7 @@
     "schemaConfigurations": [
         "recordset/add.config.json"
     ],
-    "schema": "product"
+    "schema": "recordset-product"
   },
-  "cleanup" : true,
-  "tests" : {
-    "params" : {
-      "schemaName" : "product",
-      "table_name" : "accommodation",
-      "record_count" : 5,
-      "title": "Best Western Plus Amedia Art Salzburg",
-      "website": "https://www.bestwestern.com/en_US/book/hotels-in-salzburg/best-western-plus-amedia-art-salzburg/propertyCode.89145.html",
-      "rating": "3.50",
-      "summary": "The BEST WESTERN PLUS Amedia Art Salzburg is located near the traditional old part of town, near the highway, near the train station and close to the exhibition center of Salzburg.\nBEST WESTERN PLUS Amedia Art Salzburg offers harmony of modern technique and convenient atmosphere to our national and international business guest and tourists.",
-      "opened_on": "6/15/1976, 12:00:00 AM",
-      "luxurious": "false"
-    }
-  }
+  "cleanup" : true
 }
