+++ test/e2e/specs/default-config/protractor.conf.js
@@ -4,9 +4,7 @@ var config = pConfig.getConfig({
     // This config is meant to be run as part of the parallel tests configuration
     configFileName: 'parallel-configs/default-config.dev.json',
     specs: [
-        "recordset/*.spec.js",
-        "record/*.spec.js",
-        "viewer/*.spec.js"
+        "*/*.spec.js"
     ],
     setBaseUrl: function(browser, data) {
         browser.params.url = process.env.CHAISE_BASE_URL;
