+++ test/e2e/specs/default-config/recordedit/composite-key.spec.js
@@ -1,106 +1,102 @@
-var chaisePage = require('../../../../utils/chaise.page.js');
-var recordEditHelpers = require('../../helpers.js');
+var chaisePage = require('../../../utils/chaise.page.js');
+var recordEditHelpers = require('../../../utils/recordedit-helpers.js');
+var testParams = {
+    table_name: "accommodation",
+    column_names: ["first_name", "last_name", "product-person_product-person-fk"]
+};
 
 describe('Add a record,', function() {
 
-	var params, testConfiguration = browser.params.configuration.tests, testParams = testConfiguration.params, EC = protractor.ExpectedConditions;
+    describe("For table " + testParams.table_name + ",", function() {
 
-    for (var i=0; i< testParams.tables.length; i++) {
+        var record;
 
-    	(function(tableParams, index) {
+        beforeAll(function () {
+            browser.ignoreSynchronization=true;
+            browser.get(browser.params.url + "/recordedit/#" + browser.params.catalogId + "/product-person:" + testParams.table_name);
+        });
 
-    		describe("For table " + tableParams.table_name + ",", function() {
+        describe("Presentation and validation for an entity with a composite key,", function() {
 
-    			var record;
+            var rows;
 
-				beforeAll(function () {
-					browser.ignoreSynchronization=true;
-					browser.get(browser.params.url + ":" + tableParams.table_name);
-			    });
+            it("the composite key should be filled in.", function() {
+                var EC = protractor.ExpectedConditions;
+                var modalTitle = chaisePage.recordEditPage.getModalTitle();
 
-                describe("Presentation and validation for an entity with a composite key,", function() {
+                // make sure recordedit is loaded
+                chaisePage.waitForElement(element(by.id("submit-record-button"))).then(function() {
 
-                    var rows;
+                    return chaisePage.recordEditPage.getModalPopupBtnsUsingScript();
+                }).then(function(popupBtns) {
 
-                    it("the composite key should be filled in.", function() {
-                        var modalTitle = chaisePage.recordEditPage.getModalTitle();
+                    return chaisePage.clickButton(popupBtns[3]);
+                }).then(function() {
+                    // wait for the modal to open
+                    browser.wait(EC.visibilityOf(modalTitle), browser.params.defaultTimeout);
 
-                        // make sure recordedit is loaded
-                        chaisePage.waitForElement(element(by.id("submit-record-button"))).then(function() {
+                    return modalTitle.getText();
+                }).then(function(text) {
+                    // make sure modal opened
+                    expect(text.indexOf("Choose")).toBeGreaterThan(-1);
 
-                            return chaisePage.recordEditPage.getModalPopupBtnsUsingScript();
-                        }).then(function(popupBtns) {
+                    rows = chaisePage.recordsetPage.getRows();
+                    // count is needed for clicking a random row
+                    return rows.count();
+                }).then(function(ct) {
+                    expect(ct).toBeGreaterThan(0);
 
-                            return chaisePage.clickButton(popupBtns[3]);
-                        }).then(function() {
-                            // wait for the modal to open
-                            browser.wait(EC.visibilityOf(modalTitle), browser.params.defaultTimeout);
+                    var index = Math.floor(Math.random() * ct);
+                    return rows.get(index).all(by.css(".select-action-button"));
+                }).then(function(selectButtons) {
+                    return selectButtons[0].click();
+                }).then(function() {
+                    browser.wait(EC.visibilityOf(chaisePage.recordEditPage.getFormTitle()), browser.params.defaultTimeout);
 
-                            return modalTitle.getText();
-                        }).then(function(text) {
-                            // make sure modal opened
-                            expect(text.indexOf("Choose")).toBeGreaterThan(-1);
-
-                            rows = chaisePage.recordsetPage.getRows();
-                            // count is needed for clicking a random row
-                            return rows.count();
-                        }).then(function(ct) {
-                            expect(ct).toBeGreaterThan(0);
-
-                            var index = Math.floor(Math.random() * ct);
-                            return rows.get(index).all(by.css(".select-action-button"));
-                        }).then(function(selectButtons) {
-                            return selectButtons[0].click();
-                        }).then(function() {
-                            browser.wait(EC.visibilityOf(chaisePage.recordEditPage.getFormTitle()), browser.params.defaultTimeout);
-
-                            var foreignKeyInput = chaisePage.recordEditPage.getForeignKeyInputValue("Person", 0);
-                            expect(foreignKeyInput.getAttribute("value")).toBeDefined();
-                        });
-                    });
+                    var foreignKeyInput = chaisePage.recordEditPage.getForeignKeyInputValue("Person", 0);
+                    expect(foreignKeyInput.getAttribute("value")).toBeDefined();
                 });
-
-                describe("Submit record", function() {
-					beforeAll(function() {
-						// Submit the form
-						chaisePage.recordEditPage.submitForm();
-					});
-
-					var hasErrors = false;
-
-					it("should have no errors, and should be redirected", function() {
-						chaisePage.recordEditPage.getAlertError().then(function(err) {
-							if (err) {
-								expect("Page has errors").toBe("No errors");
-								hasErrors = true;
-							} else {
-								expect(true).toBe(true);
-							}
-						});
-					});
-
-					it("should be redirected to record page with correct values.", function() {
-						if (!hasErrors) {
-
-                            // After submitting 1 record in RecordEdit, the expected record
-                            // page url will have a id of 1 because it'll always be the first
-                            // row of this table in the new catalog created by this set of composite key tests.
-                            var redirectUrl = browser.params.url.replace('/recordedit/', '/record/');
-                            redirectUrl += ":" + tableParams.table_name + '/id=1';
-
-                            chaisePage.waitForUrl(redirectUrl, browser.params.defaultTimeout).then(function() {
-                                expect(browser.driver.getCurrentUrl()).toBe(redirectUrl);
-
-                                recordEditHelpers.testRecordAppValuesAfterSubmission(tableParams);
-                            }, function() {
-                                console.log("          Timed out while waiting for the url to be the new one");
-                                expect(browser.driver.getCurrentUrl()).toBe(redirectUrl);
-                            });
-						}
-					});
-				});
-
             });
-        })(testParams.tables[i], i);
-    }
+        });
+
+        describe("Submit record", function() {
+            beforeAll(function() {
+                // Submit the form
+                chaisePage.recordEditPage.submitForm();
+            });
+
+            var hasErrors = false;
+
+            it("should have no errors, and should be redirected", function() {
+                chaisePage.recordEditPage.getAlertError().then(function(err) {
+                    if (err) {
+                        expect("Page has errors").toBe("No errors");
+                        hasErrors = true;
+                    } else {
+                        expect(true).toBe(true);
+                    }
+                });
+            });
+
+            it("should be redirected to record page with correct values.", function() {
+                if (!hasErrors) {
+
+                    // After submitting 1 record in RecordEdit, the expected record
+                    // page url will have a id of 1 because it'll always be the first
+                    // row of this table in the new catalog created by this set of composite key tests.
+                    var redirectUrl = browser.params.url + "/record/#" + browser.params.catalogId + "/product-person:" + testParams.table_name + '/id=1';
+
+                    chaisePage.waitForUrl(redirectUrl, browser.params.defaultTimeout).then(function() {
+                        expect(browser.driver.getCurrentUrl()).toBe(redirectUrl);
+
+                        recordEditHelpers.testRecordAppValuesAfterSubmission(testParams.column_names);
+                    }, function() {
+                        console.log("          Timed out while waiting for the url to be the new one");
+                        expect(browser.driver.getCurrentUrl()).toBe(redirectUrl);
+                    });
+                }
+            });
+        });
+
+    });
 });
