+++ test/e2e/specs/default-config/recordedit/submission-disabled.conf.js
@@ -1,21 +1,12 @@
-var pConfig = require('./../../../../utils/protractor.configuration.js');
+var pConfig = require('./../../../utils/protractor.configuration.js');
 
 var config = pConfig.getConfig({
-
-  // Change this to your desired filed name and Comment below testConfiguration object declaration
-    configFileName: 'recordedit-submission-disabled.dev.json',
-
-  /* Just in case if you plan on not giving a file for configuration, you can always specify a testConfiguration object
-   * Comment above 2 lines
-   * Empty configuration will run test cases against catalog 1 and default schema
-    testConfiguration: {
-      ....
-    },
-  */
-
-    page: 'recordedit',
+    configFileName: 'recordedit/submission-disabled.dev.json',
+    specs: [
+        "submission-disabled.spec.js"
+    ],
     setBaseUrl: function(browser, data) {
-      browser.params.url = process.env.CHAISE_BASE_URL + "/recordedit" + "/#" + data.catalogId + "/" + data.schema.name;
+      browser.params.url = process.env.CHAISE_BASE_URL;
       return browser.params.url;
     }
 });
