+++ test/e2e/specs/default-config/recordset/edit.spec.js
@@ -1,5 +1,6 @@
 var chaisePage = require('../../../utils/chaise.page.js');
 var testParams = {
+    schemaName: "recordset-multi-edit",
     table_name: "multi-add-table",
     default_page_limit: 25,
     limit: 10,
@@ -11,7 +12,7 @@ describe('Recordset edit records,', function() {
     describe("recordset shows results with no limit defined,", function() {
         beforeAll(function() {
             browser.ignoreSynchronization = true;
-            browser.get(browser.params.url + "/recordset/#" + browser.params.catalogId + "/multi-add:" + testParams.table_name);
+            browser.get(browser.params.url + "/recordset/#" + browser.params.catalogId + "/" + testParams.schemaName + ":" + testParams.table_name);
         });
 
         it("clicking edit will show forms based on the default page size of " + testParams.default_page_limit + ".", function() {
@@ -38,7 +39,7 @@ describe('Recordset edit records,', function() {
     describe("recordset url includes a limit,", function() {
         beforeAll(function() {
             browser.ignoreSynchronization = true;
-            browser.get(browser.params.url + "/recordset/#" + browser.params.catalogId + "/multi-add:" + testParams.table_name + "?limit=" + testParams.limit);
+            browser.get(browser.params.url + "/recordset/#" + browser.params.catalogId + "/" + testParams.schemaName + ":" + testParams.table_name + "?limit=" + testParams.limit);
         });
 
         it("clicking edit will show forms based on the limit of " + testParams.limit + " in the uri.", function() {
@@ -68,7 +69,7 @@ describe('Recordset edit records,', function() {
         });
 
         it("without a limit, clicking edit will show all forms with int=23.", function() {
-            browser.get(browser.params.url + "/recordset/#" + browser.params.catalogId + "/multi-add:" + testParams.table_name + "/int=23");
+            browser.get(browser.params.url + "/recordset/#" + browser.params.catalogId + "/" + testParams.schemaName + ":" + testParams.table_name + "/int=23");
 
             chaisePage.waitForElement(element(by.id("divRecordSet"))).then(function() {
                 return chaisePage.recordsetPage.getRows().count();
@@ -90,7 +91,7 @@ describe('Recordset edit records,', function() {
         });
 
         it("with a limit of " + testParams.limit + ", clicking edit will show all forms with int=23.", function() {
-            browser.get(browser.params.url + "/recordset/#" + browser.params.catalogId + "/multi-add:" + testParams.table_name + "/int=23?limit=" + testParams.limit);
+            browser.get(browser.params.url + "/recordset/#" + browser.params.catalogId + "/" + testParams.schemaName + ":" + testParams.table_name + "/int=23?limit=" + testParams.limit);
 
             chaisePage.waitForElement(element(by.id("divRecordSet"))).then(function() {
                 return chaisePage.recordsetPage.getRows().count();
